u-boot-kiss: add recipe to build mainline U-Boot
Using recipes-bsp/u-boot/u-boot.inc from oe-core is very handy and allows
to write a working recipe with only a few additional lines. It uses git to
fetch the sources, thus we need to set SRVREV to a commit. For clarity, add
a comment before the SRCREV line with the tag pointing to this commit.

Signed-off-by: Luca Ceresoli <luca.ceresoli@bootlin.com>