u-boot-kiss: store environment on mmc partition 4 for stm32mp1 products
We want to store the U-Boot environment as a file in the fourth partition
of the uSD card, so add a defconfig fragment to configure U-Boot in such a
way.

Note we only need to add the fragment to SRC_URI. The code from oe-core
(recipes-bsp/u-boot/u-boot-configure.inc) takes care of applying it.

Signed-off-by: Luca Ceresoli <luca.ceresoli@bootlin.com>