linux-kiss: add per-machine defconfigs
The kernel.bbclass is using the default architecture defconfig unless it is
explicitly provided a defconfig. This default works, but for many realistic
products having a customized defconfig is preferred in order to enable all
the needed features and drivers while trimming the kernel and modules size
disabling unneeded ones.

As a preliminary step, add a new defconfig for each machine. This defconfig
is identical to the default defconfig (multi_v7_defconfig for ARM32),
changes will be done in following commits. The defconfigs have been
generated as in the following example for hte dogbonedark machine:

  bitbake -c savedefconfig linux-kiss
  cp tmp-glibc/work/dogbonedark-oe-linux-gnueabi/linux-kiss/6.6.2-r0/build/defconfig \
     ../meta-kiss/recipes-kernel/linux/linux-kiss/dogbonedark/defconfig
  git add ../meta-kiss/recipes-kernel/linux/linux-kiss/dogbonedark/defconfig

Enabling them just requires adding the defconfig file to SRC_URI and
kernel.bbclass will use them automatically.

Signed-off-by: Luca Ceresoli <luca.ceresoli@bootlin.com>