sl: avoid warning due to SL license being non-common in oe-core
SL has a license listed in SPDX and ships a LICENSE file. However the SL
license is not in the openembedded-core list of common licenses, so when
building sl and kiss-image respectively the following warnings are
produced:

  WARNING: sl-1.0+gitAUTOINC+923e7d7ebc-r0 do_populate_lic: QA Issue: sl: No generic license file exists for: SL in any provider [license-exists]
  WARNING: kiss-image-1.0-r0 do_rootfs: QA Issue: The license listed SL was not in the licenses collected for recipe sl [license-file-missing]

This is not an error but rather due to the recipe being not in the common
licenses list.

Fix by using the NO_GENERIC_LICENSE variable whose goal is exactly to avoid
such warnings.

Fixes: https://github.com/bootlin/simplest-yocto-setup/issues/4
Reported-by: https://github.com/fabioestevam
Signed-off-by: Luca Ceresoli <luca.ceresoli@bootlin.com>